backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch==2.5.1
torchvision==0.20.1
numpy==1.26.2
opencv-python==4.8.1.78
mediapipe==0.10.8
safetensors==0.4.0
tqdm==4.66.1
scikit-learn==1.3.2